refactor: reorg ShipProxy.lua to Ship/
